ref: move to ssr for fetching
